moving the ingestion of a df to ingest.py
